♻️ refactor : 답변페이지 일부 및 requestApi 수정
