+++ src/hooks/useRespondAnswer.jsx
@@ -1,0 +1,32 @@
+import { useState } from 'react';
+import requestApi from '../utils/requestApi';
+
+const useRespondAnswer = ({
+  setCurrentType,
+  item,
+  setItem,
+  questionId,
+  content,
+}) => {
+  const [answer, setAnswer] = useState(content);
+  const respondAnswer = () => {
+    setCurrentType('Answer');
+    const isRejected = answer.replaceAll(/[\n\s]/g, '') === '답변거절';
+    const postData = {
+      content: answer,
+      isRejected,
+    };
+    if (!item) {
+      postData.questionId = questionId;
+      requestApi(`questions/${questionId}/answers/`, 'post', postData).then(
+        (result) => setItem(result),
+      );
+      return;
+    }
+    requestApi(`answers/${item.id}/`, 'patch', postData);
+  };
+
+  return { respondAnswer, answer, setAnswer };
+};
+
+export default useRespondAnswer;
